+++ app/build.gradle
@@ -25,7 +25,7 @@ android {
     }
 
     def API_BASE = "API_BASE"
-   applicationVariants.all { variant ->
+    applicationVariants.all { variant ->
         variant.buildConfigField "String", API_BASE, "\"http://gateway.marvel.com/\""
     }
 }
@@ -40,21 +40,22 @@ kapt {
 
 dependencies {
     implementation fileTree(dir: 'libs', include: ['*.jar'])
-    implementation"org.jetbrains.kotlin:kotlin-stdlib-jdk7:$kotlin_version"
+    implementation "org.jetbrains.kotlin:kotlin-stdlib-jdk7:$kotlin_version"
     implementation 'androidx.appcompat:appcompat:1.1.0'
     implementation 'androidx.core:core-ktx:1.1.0'
     implementation 'androidx.constraintlayout:constraintlayout:1.1.3'
 
     implementation 'com.google.android.material:material:1.0.0'
 
-    implementation 'androidx.lifecycle:lifecycle-extensions:2.2.0-rc01'
+    implementation 'androidx.lifecycle:lifecycle-extensions:2.2.0-rc02'
 
     implementation 'de.hdodenhof:circleimageview:3.0.0'
 
     implementation 'com.airbnb.android:epoxy:3.4.2'
+    implementation 'androidx.legacy:legacy-support-v4:1.0.0'
     kapt "com.airbnb.android:epoxy-processor:3.4.2"
 
-    implementation 'androidx.recyclerview:recyclerview:1.0.0'
+    implementation 'androidx.recyclerview:recyclerview:1.1.0'
 
     implementation 'com.github.bumptech.glide:glide:4.9.0'
     kapt 'com.github.bumptech.glide:compiler:4.9.0'
@@ -80,9 +81,9 @@ dependencies {
     androidTestImplementation 'androidx.test:runner:1.2.0'
     androidTestImplementation 'androidx.test.espresso:espresso-core:3.2.0'
 
-    testImplementation 'org.mockito:mockito-core:2.28.2'
+    testImplementation 'org.mockito:mockito-core:3.1.0'
     testImplementation 'androidx.arch.core:core-testing:2.1.0'
     testImplementation 'com.squareup.okhttp3:mockwebserver:4.2.1'
     testImplementation 'org.koin:koin-test:2.0.1'
-    androidTestImplementation 'org.mockito:mockito-android:2.28.2'
+    androidTestImplementation 'org.mockito:mockito-android:3.1.0'
 }
